# backend/src/main/resources/static/openapi/punenest-api.yaml
openapi: 3.1.0
info:
  title: PuneNest API
  version: "1.3.0"
  summary: Pune-first real-estate marketplace backend API.
  description: |
    Enterprise API contract for **PuneNest**, a Pune-first real-estate marketplace
    (broker-optional listings, rentals, deals, services and back-office operations).

    **This specification is the single source of truth for the PuneNest HTTP API.**
    Endpoints, request/response shapes, and enums are defined here and nowhere else.
    Companion docs describe *behaviour and data*, not the wire contract:
      - `docs/system/data-model.md`     â€” ER map / DB persistence design (field shapes are the schemas below)
      - `docs/system/cross-cutting.md`  â€” auth, contact gate, maker-checker, audit rules
      - `docs/system/platform-architecture.md` â€” component/vendor decisions & ADRs (KYC, payments, verification posture)
      - `docs/system/backend-api-architecture-review.md` â€” architecture-review narrative
      - `docs/flows/**`                 â€” per-feature behavioural specs

    ### Changelog
    - 1.3.0 — **Flatmates re-modelled around seeker intent.** The flatmates/rooms/groups triple is
      replaced by two feeds, `move-in` and `team-up` (`FlatmateTab`; the old values survive as
      deprecated read aliases). `FlatmatePost` is renamed `FlatmateSeekerPost` and re-shaped to what
      a seeker actually posts; `FlatmateRoom`, `FlatmateGroup`, `FlatmateRequest`, `FlatmateReview`,
...
      "verified contact only".
...

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /properties/{id}/status:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    patch:
      tags: [Moderation]
      summary: Approve / reject listing
      operationId: setPropertyStatus
      # S28: Approving/rejecting a listing is the core moderation task.
      x-roles: [staff, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyStatusUpdate' }
      responses:
        '200': { description: Updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /properties/{id}/toggle-featured:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    post:
      tags: [Moderation]
      summary: Toggle featured flag
      operationId: toggleFeatured
      # S28: Merchandising a listing on the homepage.
      x-roles: [staff, admin]
      responses:
        '200': { description: Toggled }
        '404': { $ref: '#/components/responses/NotFound' }
  /properties/{id}/flag:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    post:
      tags: [Moderation]
      summary: Flag listing
      operationId: flagProperty
      # S28: Raising a moderation flag.
      x-roles: [staff, admin]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { description: Flagged }
    delete:
      tags: [Moderation]
      summary: Clear flag
      operationId: clearFlag
      # S28: Clearing a moderation flag.
      x-roles: [staff, admin]
      responses:
        '204': { description: Cleared }
  /properties/{id}/archive:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    patch:
      tags: [Moderation]
      summary: Archive listing (soft-delete)
      operationId: archiveProperty
      # S28: deliberately NO x-roles. DUAL AUDIENCE: the owner soft-deleting their own listing, or staff/admin taking one down. Role-gating this would remove an owner's ability to delete their own listing. Guard is owner-or-staff, enforced server-side.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { description: Archived }
        '404': { $ref: '#/components/responses/NotFound' }
  /properties/{id}/restore:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    patch:
      tags: [Moderation]
      summary: Restore archived listing
      operationId: restoreProperty
      # S28: deliberately NO x-roles. DUAL AUDIENCE: see archiveProperty.
      responses:
        '200': { description: Restored (status reset to pending) }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /properties/{id}/admin:
    parameters: [ { $ref: '#/components/parameters/PropertyId' } ]
    patch:
      tags: [Moderation]
      summary: Admin update of property fields
      description: >
        Admin field-level update. NOTE â€” the source contract reuses
        `PATCH /properties/{id}`; disambiguated here from the public getter to
        avoid method/verb collision (documented assumption).
      operationId: adminUpdateProperty
      # S28: Editing another user's listing to correct it.
      x-roles: [staff, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListingUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        # Slice-15 spec fix S67. The one path that edits a listing the caller does not own had
        # no failure documented at all, so "id does not exist" was undefined behaviour.
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /admin/properties:
    get:
      tags: [Moderation]
      summary: Listing moderation queue
      operationId: listPropertiesForModeration
      description: >
        Listings of every status including archived, newest first. The moderation writes
        (`setPropertyStatus`, `toggleFeatured`, `flagProperty`, `clearFlag`, `adminUpdateProperty`)
        all address a listing by id and shipped without any read that could produce one for an
        unapproved listing: `searchProperties` pins `status='approved' AND archived=false`
        unconditionally — it takes no principal, so it cannot relax for staff — and `myListings` is
        scoped to the caller's own account. Unlike `searchProperties`, `status` here *widens*.
        Owner contact is masked, as on every ops surface.
      x-roles: [staff, admin]
      parameters:
        - { name: deal, in: query, schema: { $ref: '#/components/schemas/DealIntent' } }
        - { name: type, in: query, schema: { type: string } }
        - { name: locality, in: query, schema: { type: string }, description: 'Locality *slug*, as on searchProperties' }
        - { name: bhk, in: query, schema: { type: integer, minimum: 0 } }
        - { name: minPrice, in: query, schema: { type: integer, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: integer, minimum: 0 } }
        - { name: furnishing, in: query, schema: { $ref: '#/components/schemas/Furnishing' } }
        - { name: possession, in: query, schema: { $ref: '#/components/schemas/PropertyPossession' } }
        - { name: q, in: query, description: Free-text over title + locality, schema: { type: string } }
        - name: status
          in: query
          description: Narrow to one moderation status. Omit for every status.
          schema: { $ref: '#/components/schemas/PropertyStatus' }
        - name: archived
          in: query
          description: >
            Tri-state. Omit for both live and archived listings; `true` for archived only; `false`
            for live only. A separate axis from `status` — archiving preserves the moderation state
            the listing was archived from.
          schema: { type: boolean }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
      responses:
        '200':
          description: Paged listings, every status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/Property' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /me/contact-requests:
    get:
      tags: [Leads & Contact]
      summary: View incoming contact requests (owner)
      operationId: myContactRequests
      description: >
        Newest first, owner-scoped. Paged: this collection grows with inbound demand rather than with
        the owner's own actions, so a popular listing is exactly what makes it large. Use
        `myPendingContactCount` for the "waiting on you" badge - counting one page under-reports.
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Paged requests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/ContactRequest' }
...
  /me/contact-requests/{reqId}:
    parameters: [ { $ref: '#/components/parameters/ReqId' } ]
    patch:
      tags: [Leads & Contact]
      summary: Approve / decline contact request
      operationId: respondContactRequest
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusUpdate' }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
  # Slice-12 spec fix S45. `GET /enquiries` and the `Enquiry` schema were removed here,
  # executing the sunset this document wrote for itself in slice 4: "NOT IMPLEMENTED and will not
  # be. Sunset: remove from this spec once the admin lead views land." Those views shipped in
  # slice 8. Nothing in the platform writes an `enquiries` row -- the contact-request surface
  # (ADR-019) replaced it before either was implemented -- so the operation could only ever have
  # returned an empty page, which is a worse admin surface than none. V22 drops the table.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by the Identity & Access domain. Send as
        `Authorization: Bearer <token>`. Short-lived (15 min); refresh via
        `/auth/refresh`. Claims carry `sub`, `role`, `mobileVerified`, and
        `aadhaarVerified` (an opt-in trust **badge**, not a gate â€” participation only
        needs `mobileVerified`); `staff` tokens also carry `team`.
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.punenest.in/.well-known/openid-configuration
      description: >
        Target-state federation. The modular monolith issues its own JWTs today;
        an external OIDC provider (Keycloak/Auth0) can be adopted without changing
        resource-server contracts.
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index.
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: Page size (max 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      description: '`field,(asc|desc)`; repeatable for multi-sort.'
      schema: { type: string, example: createdAt,desc }
    PropertyId:
      name: id
      in: path
      required: true
      description: Property slug or id.
      schema: { type: string, example: 3bhk-koregaon-park-xf12 }
...
    ReqId:
      name: reqId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Forbidden:
      description: Authenticated but not permitted (RBAC, or owner accepts verified contacts only).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: forbidden, message: You do not have permission to perform this action, status: 403 }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: not_found, message: Property not found, status: 404 }
...
    ValidationError:
      description: Semantic validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
  schemas:
...
    Error:
      type: object
      required: [error, message, status]
      properties:
        error: { type: string, description: Machine-readable code., example: not_found }
        message: { type: string, description: Human-readable detail. }
        status: { type: integer, description: HTTP status., example: 404 }
        traceId: { type: string, description: Correlation id for observability. }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            fields:
              type: array
              items:
                type: object
                properties:
                  field: { type: string, example: mobile }
                  message: { type: string, example: must be a 10-digit Indian mobile }
    PageEnvelope:
      type: object
      description: Envelope for paginated collections; `content` is supplied per endpoint.
      properties:
        page: { type: integer, example: 0 }
        size: { type: integer, example: 20 }
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        sort: { type: string, example: createdAt,desc }
    Mobile:
      type: string
      pattern: '^[6-9][0-9]{9}$'
      example: '9876543210'
    Money:
      type: integer
      format: int64
      description: Amount in whole Indian rupees (INR). No decimals/paise.
      example: 2500000
    DealIntent:
      type: string
      description: Transaction intent.
      enum: [buy, rent]
    Furnishing:
      type: string
      enum: [unfurnished, semi-furnished, furnished]
    PropertyPossession:
      type: string
      description: >
        When the property can be occupied. Absent/null means "not stated", which is
        deliberately distinct from all three values â€” an unrecorded possession must not
        satisfy a `ready-to-move` search. Land and plot listings have no meaningful
        possession state and stay unset.
      enum: [ready-to-move, new-launch, under-construction]
    PropertyStatus:
      type: string
      enum: [pending, approved, rejected, flagged, archived]
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        note: { type: string, maxLength: 500 }
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
...
    Property:
      allOf:
        - $ref: '#/components/schemas/PropertySummary'
        - type: object
          properties:
            description: { type: string }
            deposit: { $ref: '#/components/schemas/Money' }
            maintenance: { $ref: '#/components/schemas/Money' }
            negotiable: { type: boolean }
            reraId: { type: string }
            carpetArea: { type: number }
            builtUpArea: { type: number }
            superBuiltUpArea: { type: number }
            floor: { type: integer }
            totalFloors: { type: integer }
            facing: { type: string }
            amenities:
              type: array
              items: { type: string }
            images:
              type: array
              items: { type: string, format: uri }
            floorPlan: { type: string, format: uri }
            video: { type: string, format: uri }
            address: { type: string }
            pincode: { type: string }
            views: { type: integer, description: Denormalized view counter }
            enquiries: { type: integer, description: Denormalized enquiry counter }
            featured: { type: boolean }
            flagReason: { type: [string, 'null'], description: Admin moderation flag; empty when unflagged }
            archived: { type: boolean, description: Soft-delete flag. A separate axis from status - always false on the public detail read; meaningful on /me/listings and /admin/properties. }
            ownerVerified: { type: boolean }
            ownershipVerified: { type: boolean }
            societyVerified: { type: boolean }
            conveyanceDone: { type: boolean }
            docsCount: { type: integer }
            owner:
              type: object
              description: Owner summary; contact number masked until the contact-gate is passed.
              properties:
                id: { type: string }
                name: { type: string }
                mobile: { type: string, description: Masked (e.g. 98XXXXX210) until unlocked }
                verified: { type: boolean }
            adminPipeline:
              type: object
              description: Admin post-on-behalf onboarding pipeline (staff-created listings).
              properties:
                postedByAdmin: { type: boolean }
                postedByStaff: { type: [string, 'null'] }
                pipelineStage: { type: string, enum: [listed, docs_submitted, photos_uploaded, aadhaar_verified, claim_sent, claimed] }
                claimLinkSent: { type: boolean }
                photosUploaded: { type: boolean }
                aadhaarVerified: { type: boolean }
                reminderCount: { type: integer }
...
    ListingCreate:
      type: object
      required: [title, deal, propertyType, price, locality, city]
      properties:
        title: { type: string }
        deal: { $ref: '#/components/schemas/DealIntent' }
        propertyType: { type: string }
        bhk: { type: number }
        price: { $ref: '#/components/schemas/Money' }
        deposit: { $ref: '#/components/schemas/Money' }
        maintenance: { $ref: '#/components/schemas/Money' }
        negotiable: { type: boolean }
        area: { type: number }
        areaUnit: { type: string, default: sqft }
        furnishing: { $ref: '#/components/schemas/Furnishing' }
        locality: { type: string }
        city: { type: string }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        reraId: { type: string }
        possession: { $ref: '#/components/schemas/PropertyPossession' }
        amenities: { type: array, items: { type: string } }
        images: { type: array, items: { type: string, format: uri } }
        description: { type: string }
    ListingUpdate:
      allOf:
        - $ref: '#/components/schemas/ListingCreate'
        - type: object
          description: All fields optional on update (PATCH semantics).
    PropertyStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          # Deliberately NOT $ref PropertyStatus. That enum has five values; this operation accepts
          # three. `flagged` is reached through flagProperty (which also records a reason) and
          # `archived` through archiveProperty (a separate axis, and owner-or-staff rather than
          # staff-only), so allowing either here would be a second, reason-less way to do something
          # the API already models properly. The implementation 400s on the other two; before this
          # was narrowed the contract promised five and delivered three.
          enum: [pending, approved, rejected]
          description: Moderation decision. Use flagProperty / archiveProperty for the other states.
        reason: { type: string, description: Recorded on the audit row; required by convention when rejecting }
...
    ContactRequest:
      type: object
      properties:
        id: { type: string }
        propertyId: { type: string }
        requester: { $ref: '#/components/schemas/Party' }
        status: { type: string, enum: [pending, approved, declined] }
        contact:
          type: [object, 'null']
          description: Present only after the gate is passed & owner grants.
          properties:
            name: { type: string }
            mobile: { $ref: '#/components/schemas/Mobile' }
        createdAt: { type: string, format: date-time }
    # Slice-12 spec fix S45: `Enquiry` removed with its only operation. See /enquiries above.
...